refactor XliteratorNew => Xliterator
